local/requirements.txt
# VisaWire Local Voice Agent - Dependencies
# Core
fastapi==0.115.0
uvicorn==0.30.0
python-dotenv==1.0.1
websockets==13.1
httpx==0.27.0

# Database
psycopg[binary]
redis==5.0.8

# ML / Models
torch>=2.1.0
transformers>=4.57.0
accelerate>=0.27.0
soundfile==0.12.1
numpy>=1.24.0

# ASR — Pingala V1
pingala-shunya>=0.1.5

# TTS — English (Qwen3-TTS)
qwen-tts>=0.1.1

# TTS — Hindi (Parler-TTS)
parler-tts>=0.2.3

# TTS — Fallback (Windows SAPI5 offline)
pyttsx3>=2.90

# Testing
pytest==8.3.0
pytest-asyncio==0.24.0

# Audio processing
librosa==0.10.2
scipy>=1.12.0
